Migrates r-research dependencies to r-data (Recidiviz/recidiviz-data#20869)
## Description of the change

This PR migrates `recidiviz-research` dependencies to `recidiviz-data`.
As discussed in eng leads, the value of this change is reducing the
amount of upkeep required in `r-research`. An analyst would `pipenv
shell` from `r-data`, then change directories to `r-research`, thereby
using the latter repo _only_ for storage of notebooks and code reviews
but using the former for package management/CI.

Note that I only migrated python packages to the `r-data` pipfile that
are imported somewhere in `r-research`. It's possible I missed packages
that are imported but not in the pipfile (e.g. are dependencies of
packages not migrated). The fix to this, when someone encounters a
package that is not installed is to add the relevant package to the
_`r-data`_ pipfile.

I tested this change locally on a couple of notebooks and got expected
results!

Still todo:

1. I need to move over a few more python files
2. Test nbautoexport
3. I'd like to deprecate the pipfile and other files in `r-research` so
that it only contains notebooks and files exported from notebooks

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#20861

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: 547ffb9373b4e39b37742ed88685ea02ab5d5bbd