[asset generation] new unit chart (Recidiviz/recidiviz-data#21260)
Adds a new chart for Outliers unit data that will ultimately replace and
deprecate the existing chart. It's quite a bit more complex than the old
one, since it has to create a "swarm" layout with the statewide dataset
in addition to plotting the caseloads in the unit (the latter part is
pretty similar to the previous chart and uses the same data). There are
some public packages out there for this sort of thing, but none that
quite meet our layout requirements, so I opted for a custom
implementation.

GitOrigin-RevId: 47d0352e89aa5aedff42892203436f6973ca4057