[US_PA] Add logic to include more units in location metadata  (Recidiviz/recidiviz-data#21347)
## Description of the change

The `us_pa_location_metadata` reference view query automatically set
`supervision_unit_* `fields to `NULL` if the name of the location didn't
include "UNIT", but some location names in PA use "UNT" instead of the
full word "UNIT". This change tweaks the logic in the reference view
query so that those units get included in location metadata.

The goal of this change is to improve hydration in `location_metadata`
so we see fewer NULLs when trying to associate officer’s unit
assignments with the unit assignment of their current supervisors.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 98e2ce27143acc1d96d3323ba8f8d5bfaa2fcac5