Log gunicorn errors to log file (Recidiviz/recidiviz-data#21439)
## Description of the change

It seems like `gunicorn` is terminating our workers. Checking the docker
logs for app engine, I can see the state bq refresh processing, then all
of the sudden the logs stop and there are no more logs for that worker
pid.

We try to forward `gunicorn.error` logs to StackDriver via structured
logging, but that doesn't work if gunicorn terminates the worker before
it can be sent to GCP.

This PR should give us some more visibility into what might be going on.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to ongoing refresh issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: de9631a03df34c109577dd04b555c1cad8194108