[Workflows] Create handle_send_sms endpoint (Recidiviz/recidiviz-data#21371)
## Description of the change

Creates an endpoint, `workflows/external_request/handle_send_sms`, which
enqueues an sms or mms message to be sent via the twilio api. This
endpoint will ultimately be called by a cloud task queue to facilitate
things like retrying. It will not be called by external clients.

Not covered in this PR:
- Verifying the user has permissions beyond access to call our backend
- Writing status to firestore
- Setting up a callback to get delivery status
- Retries

There is no vulnerability in deploying this as the twilio secrets are no
available in prod. Additionally, even if they were, we are on a trial
account which only allows you to message pre-verified numbers which is
currently limited to two employees.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20444

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: 2060959983adc048769c7d4f9417ccbc6d62b8a7