[US_TN] Hydrating StateStaffRolePeriod in TN (Recidiviz/recidiviz-data#21431)
## Description of the change

This PR hydrates StateStaffRolePeriod in TN. Some important things to
note.

- We use the Staff raw data table to do this, but this table is updated
in place by TN, so we query Staff@ALL to construct periods over time.
- To generate periods before the current snapshot we have, we include an
arbitrary start of 1900-01-01 for all people currently in the ALL table.
- For anyone that has more than one StaffTitle in the Staff table at
this time (very few because of the pruning that we've been doing), we
assign them the StaffTitle of the FIRST title they had in the table
before it changed.
- We set the always_historical_flag to False in raw data to ensure this
file no longer gets pruned in order to construct more accurate periods
over time.
- We remove periods where someone is inactive with no end date, since
this would imply they are not an active employee at the moment. However,
we do not remove periods of inactivity with end dates since that implies
they went from active to inactive and back to active and we want to
capture all of these periods.
- of all state_agents currently ingested with open cases, all but 2
StaffIDs show up in these StateStaffRolePeriods as active agents.
However, the 2 that are not showing up are not present in the raw data,
signifying perhaps an anomaly with pruning and a rogue old entity or
switch assignment that was never cleaned up by TN.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18023

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 64c6bf776444a816816dee3ed657d11bae80f86b