[Justice Counts] handle single-page bulk upload (Recidiviz/recidiviz-data#21236)
## Description of the change

This PR provides functionality for a user to upload a single page bulk
upload sheet that contains data for multiple metrics. This requires that
the sheet contains a `metric` column that identifies which metric a
given row is associated with.

This also assumes that the proper cells are empty:

- a row that contains data for an aggregate metric has empty cells in
the `breakdown_category` and `breakdown` columns.
- a row that contains data for an annual metric has an empty `month`
cell

Another assumption: the `breakdown_category` is the same as that
metric's `metricfile.disaggregation_column_name`

An example test file for a single page bulk upload with multiple metrics
can be found
[here](https://docs.google.com/spreadsheets/d/119wO8U7Jmx4HNcLnswksBafHpr265bjo/edit#gid=1983180778).

Other example templates can be found
[here](https://drive.google.com/drive/u/0/folders/1dbolg2hqriYsxOrNKSVfBgMQaZvFrdOy).

https://www.loom.com/share/c0a1365d121c4046ab396af3a36a9af5

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#20357

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
Co-authored-by: Michelle Orden <michelleorden@Michelles-MBP.hsd1.il.comcast.net>
Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
GitOrigin-RevId: eabe2935233fc08d04f3b1bdfeb9b44f5041a4a1