Add New Temporary Script to Copy Least and Most Recent Files From Primary Storage to Secondary Ingest (Recidiviz/recidiviz-data#21427)
## Description of the change

See this
[document](https://paper.dropbox.com/doc/Entity-Deletion-Automated-Raw-Data-Pruning--B5vLZ6O19TEeLkKYZcDJ7jXjAg-2Y1aqw6jiUcJdhuyyCFwk)
for more context.

Add a new temporary script that identifies on GCS the most and least
recent files of each given table on GCS, and copy them from primary
storage to secondary ingest.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 55bf7df0f3ae4be28b6915090c65e03dc59d8c2d