Adds metric tables encompassing entire JII population (Recidiviz/recidiviz-data#20989)
## Description of the change

This PR adds a new `MetricPopulationType` that encompasses all
individuals actively under the jurisdiction of a state (called
`JUSTICE_INVOLVED`. Such a population is specifically useful for
observing the metric `Person-Days: Weighted Justice Impact`. I added
this metric in Recidiviz/recidiviz-data#20794 but did not actually add it to an aggregated
metrics table.

To accommodate the above, this PR also:

1. Adds a metric for `assignments` so that it can be excluded from
populations where it is not required
2. Adjusts `generate_aggregated_metrics_view_builder()` to only include
metrics views that contain at least one hydrated metric, which allows us
to skip building views that do not contain any metrics
3. Makes some code iterative across metric types where logic could be
condensed

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f0639ec2b2a7c12316bd75fffad51a2329475909