[US_TN] Correcting DeathDate logic in IP view to properly end periods when DeathDate is reported (Recidiviz/recidiviz-data#21465)
## Description of the change

@dallen5 caught a situation where we had not truncated a period for
someone who was reported as deceased. Upon further inspection of the
view, we were not carrying DeathDate values through to any segregation
periods, causing the periods to remain open. Additionally, we were not
assigned an EndReason for periods that were truncated by the death date,
so even when periods ended on the correct dates ,we were not capturing
why.

This PR makes sure all periods in the view carry the DeathDate value
through, truncates the final period if needed when a DeathDate is
present, and assigns these periods the release reason of DEATH-DEATH
which maps to the enum value DEATH for release_reason.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cfcf58131d78c00eb01e5fa887f1128f3fa35f62