TN CR Offense Criteria (Recidiviz/recidiviz-data#21057)
## Description of the change

This PR creates the offense-related criteria views for TN compliant
reporting. In addition to the criteria queries, there are a few upstream
changes that are made in order to simplify the logic in criteria queries
themselves. This includes the following:

1. Create 4 boolean flags for offense types in TN sentences preprocessed
that are then carried through to the state agnostic sentences
preprocessed view. These are flags for `is_violent_domestic`, `is_dui`,
`is_vicitm_under_18`, and `is_homicide`. These fields are not currently
hydrated for any other state.

2. Reformat the offense description column for TN sentences as part of
the preprocessing file as opposed to part of the criteria queries

3. Create a new view `us_tn_prior_record_preprocessed` - this pulls from
the `PriorRecord_latest` raw data table and creates flags for the 6
relevant offense types (the 4 listed above + `is_violent` and
`is_sex_offense`.

The table below shows the 6 new criteria that are written as part of
this PR. The first 3 apply to both the eligible and eligible without
discretion opportunities, and the last 3 apply to the eligible without
discretion opportunity.

| Criteria    | Description | Reasons JSON | Discretionary
| --------- | ------- |-------|-------|
| No DUI Offense in Past 5 Years | Marks ineligibility for the 5 years
following a DUI offense (from both sentences and prior record) | Date of
the most recent DUI offense | No
| No Murder Convictions | Marks ineligibility for all time following a
homicide offense (from both sentences and prior record) | Date of the
most recent homicide offense | No
| Not Serving Ineligible CR Offense | Marks ineligibility for all time
when a person is actively serving a sentence of the 4 types | Array of
offense descriptions and array of projected completion dates | No
| No Prior Record with Ineligible Offenses | Marks ineligibility for all
time following the offense date of a prior record sentence for one of
the 4 types of ineligible sentences | Array of offense descriptions and
array of offense dates | Yes
| Not serving unknown CR offense | Marks ineligibility for all time when
a person is actively serving a sentence that we are not certain is not
an ineligible offense. This includes: (1) ISC sentences that are maybe
assaultive, (2) ISC sentences that are not a known offense description,
and (3) any sentence with a missing description | Array of offense
descriptions and array of projected completion dates | Yes
| Ineligible offenses expired | Marks ineligibility for the time 10
years after completing a sentence for an ineligible CR offense | Array
of offense descriptions and array of sentence completion dates | Yes

The following is the command I used to test the output (creates the 6
new criteria views + `us_tn_sentences_preprocessed`,
`sentences_preprocessed`, and `us_tn_prior_record_preprocessed`)

```
python -m recidiviz.tools.load_views_to_sandbox \
        --sandbox_dataset_prefix agaidus manual \
        --view_ids_to_load sessions.us_tn_sentences_preprocessed,sessions.sentences_preprocessed,task_eligibility_criteria_us_tn.not_serving_ineligible_cr_offense,task_eligibility_criteria_us_tn.no_dui_offense_in_past_5_years,analyst_data.us_tn_prior_record_preprocessed,task_eligibility_criteria_us_tn.not_serving_unknown_cr_offense,task_eligibility_criteria_us_tn.no_murder_convictions,task_eligibility_criteria_us_tn.no_prior_record_with_ineligible_cr_offense,task_eligibility_criteria_us_tn.ineligible_offenses_expired\
        --update_ancestors False  --update_descendants False
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#18175

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e33fcf3e3296f265975052fb42781f2ed5fdc68a